+++ src/OpenLoco/Localisation/StringIds.h
@@ -1828,4 +1828,5 @@ namespace OpenLoco::StringIds
     constexpr string_id vehicle_details_tooltip_length = 2262;
     constexpr string_id num_characters_left_int_int = 2263;
     constexpr string_id menu_giant_screenshot = 2264;
+    constexpr string_id object_selection_filename = 2265;
 }