+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 22.10+ (???)
 ------------------------------------------------------------------------
+- Feature: [#1696] Display filename in object selection window.
 - Fix: [#1680] Title music plays in scenario editor.
 
 22.10 (2022-10-09)